Logging: cast enums to int for fmt 12 format calls
fmt 12 no longer implicitly formats enums. Cast the enum arguments in the corner-text widget and three MafiaMP client log/format calls (menu event, district id, entity type) to int, matching the prior fmt 8 behavior.

+++ code/framework/src/external/imgui/widgets/corner_text.h
@@ -37,7 +37,7 @@ namespace Framework::External::ImGUI::Widgets {
         window_flags |= ImGuiWindowFlags_NoMove;
         ImGui::SetNextWindowBgAlpha(0.35f);
 
-        const auto windowName = fmt::format("Overlay #{}", corner);
+        const auto windowName = fmt::format("Overlay #{}", static_cast<int>(corner));
 
         if (ImGui::Begin(windowName.c_str(), nullptr, window_flags)) {
             ImGui::Text("%s", text.c_str());
